// Client/src/app/Models/Vehicle.ts
export class Vehicle {
    constructor() {
        this.VehicleTypeID = 0;
    }
    VehicleID: number;
    VehicleTypeID: number;
    VehicleTypeName:string;
    Name: string;
    RegistrationNo: string;
    EngineNumber: string;
    ChasisNumber: string;
    Capacity: number;
    OtherInfo: string;
    Color: string;
}